+++ pkg/analyzer/lib/dart/element/element.dart
@@ -1910,6 +1910,7 @@ abstract class LibraryElement
   FunctionElement? get entryPoint;
 
   /// The libraries that are exported from this library.
+  @Deprecated('Use CompilationUnitElement.libraryExports')
   List<LibraryElement> get exportedLibraries;
 
   /// The export [Namespace] of this library.
@@ -1926,6 +1927,7 @@ abstract class LibraryElement
   List<LibraryElement> get importedLibraries;
 
   /// Whether the library is an application that can be run in the browser.
+  @Deprecated('Not used anymore')
   bool get isBrowserApplication;
 
   /// Whether the library is the `dart:async` library.