+++ pkg/analyzer_cli/lib/src/analyzer_impl.dart
@@ -70,11 +70,16 @@ class AnalyzerImpl {
     for (final unitElement in library.units) {
       addCompilationUnitSource(unitElement, units);
     }
-    // Add referenced libraries.
+    // Add imported libraries.
     for (var child in library.importedLibraries) {
       addLibrarySources(child, libraries, units);
     }
-    for (var child in library.exportedLibraries) {
+    // Add exported libraries.
+    var exportedLibraries = library.units
+        .expand((unit) => unit.libraryExports)
+        .map((export) => export.exportedLibrary)
+        .nonNulls;
+    for (var child in exportedLibraries) {
       addLibrarySources(child, libraries, units);
     }
   }